+++ src/ethereum/frontier/eth_types.py
@@ -15,6 +15,9 @@ Types re-used throughout the specification, which are specific to Ethereum.
 from dataclasses import dataclass
 from typing import Tuple, Union
 
+from ethereum import rlp
+from ethereum.crypto import keccak256
+
 from ..base_types import (
     U256,
     Bytes,
@@ -73,6 +76,23 @@ EMPTY_ACCOUNT = Account(
     balance=U256(0),
     code=bytearray(),
 )
+
+
+def encode_account(raw_account_data: Account, storage_root: Bytes) -> Bytes:
+    """
+    Encode `Account` dataclass.
+
+    Storage is not stored in the `Account` dataclass, so `Accounts` cannot be
+    enocoded with providing a storage root.
+    """
+    return rlp.encode(
+        (
+            raw_account_data.nonce,
+            raw_account_data.balance,
+            storage_root,
+            keccak256(raw_account_data.code),
+        )
+    )
 
 
 @slotted_freezable
